+++ src/main/frontend/src/app/mapping/import-csv-dialog/import-csv-dialog.component.ts
@@ -16,42 +16,31 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-import { Component, Inject } from '@angular/core';
+import { Component, EventEmitter, Inject, Input, Output } from '@angular/core';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
-import { ApiService } from '../api.service';
+import { ApiService, ImportedMapping } from '../api.service';
 import { firstValueFrom } from 'rxjs';
 
 @Component({
   selector: 'import-csv-dialog',
   templateUrl: './import-csv-dialog.component.html',
-  styleUrls: ['./import-csv-dialog.component.scss']
+  styleUrls: ['./import-csv-dialog.component.scss'],
 })
 export class ImportCsvDialogComponent {
-  csvImportFile : File | null = null;
-  mappingName : string = "";
+  @Input({required: true}) ignoreTermTypes!: string[];
+  @Input() noWarning: boolean = false;
+  @Output() mapping = new EventEmitter<ImportedMapping>();
+  csvImportFile: File | null = null;
   constructor(
-    private api : ApiService,
-    private dialogRef : MatDialogRef<ImportCsvDialogComponent>,
-    @Inject(MAT_DIALOG_DATA) public data : {
-      ignoreTermTypes: string[],
-      noWarning: boolean,
-    },
-  ) { }
+    private api: ApiService,
+  ) {}
 
-  handleCsvFileInput(event : Event) {
+  handleCsvFileInput(event: Event) {
     const input = event.target as HTMLInputElement;
     this.csvImportFile = null;
     if (input.files) {
       if (input.files.length == 1) {
         this.csvImportFile = input.files[0];
-        if (this.mappingName == "") {
-          let ix = this.csvImportFile.name.lastIndexOf('.');
-          if (ix == -1) {
-            this.mappingName = this.csvImportFile.name;
-          } else {
-            this.mappingName = this.csvImportFile.name.slice(0, ix);
-          }
-        }
       }
     }
   }
@@ -60,24 +49,30 @@ export class ImportCsvDialogComponent {
     this.csvImportFile = null;
   }
 
-  async importCsv(mappingName : string, file : File, format : string) {
+  async importCsv(file: File, format: string) {
     try {
-      let imported = await firstValueFrom(this.api.importCsv(file, [], format, this.data.ignoreTermTypes));
+      let imported = await firstValueFrom(
+        this.api.importCsv(file, [], format, this.ignoreTermTypes)
+      );
       if (imported.warnings.length) {
-        let msg = "There were problems with the import: " +
-          imported.warnings.map(s => `${s}. `).join("");
-        if (this.data.noWarning) {
+        let msg =
+          'There were problems with the import: ' +
+          imported.warnings.map((s) => `${s}. `).join('');
+        if (this.noWarning) {
           imported.warning = msg;
-        } else if (!confirm(msg + " Continue?")) {
+        } else if (!confirm(msg + ' Continue?')) {
           return;
         }
       }
-      imported.mappingName = mappingName;
       this.unsetCsvImportFile();
-      this.dialogRef.close(imported);
+      this.mapping.emit(imported);
     } catch (err) {
-      console.error("Could not import codelist", err);
-      let msg = typeof(err) == "string" ? err : ((err as any).error ?? "Could not import codelist: unknown error (see console)");
+      console.error('Could not import codelist', err);
+      let msg =
+        typeof err == 'string'
+          ? err
+          : (err as any).error ??
+            'Could not import codelist: unknown error (see console)';
       alert(msg);
     }
   }