+++ src/main/frontend/src/app/mapping/vocabularies/vocabularies.component.ts
@@ -16,7 +16,14 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-import { Input, Output, Component, SimpleChanges, EventEmitter, ViewChild } from '@angular/core';
+import {
+  Input,
+  Output,
+  Component,
+  SimpleChanges,
+  EventEmitter,
+  ViewChild,
+} from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { map } from 'rxjs';
 import { VocabulariesDialogComponent } from '../vocabularies-dialog/vocabularies-dialog.component';
@@ -29,20 +36,18 @@ import * as ops from '../mapping-ops';
 @Component({
   selector: 'vocabularies',
   templateUrl: './vocabularies.component.html',
-  styleUrls: ['./vocabularies.component.scss']
+  styleUrls: ['./vocabularies.component.scss'],
 })
 export class VocabulariesComponent {
-  @Input() mapping : Mapping | null = null;
+  @Input() mapping: Mapping | null = null;
   @Output() run = new EventEmitter<ops.Operation>();
-  @Input() userCanEdit : boolean = false;
-  @ViewChild('table') table! : VocabulariesTableComponent;
-  vocabularies : Vocabulary[] = [];
+  @Input() userCanEdit: boolean = false;
+  @ViewChild('table') table!: VocabulariesTableComponent;
+  vocabularies: Vocabulary[] = [];
 
-  constructor(
-    private dialog : MatDialog,
-    private api : ApiService) { }
+  constructor(private dialog: MatDialog, private api: ApiService) {}
 
-  ngOnChanges(changes : SimpleChanges) {
+  ngOnChanges(changes: SimpleChanges) {
     if (this.mapping == null) {
       this.vocabularies = [];
     } else {
@@ -51,49 +56,65 @@ export class VocabulariesComponent {
     }
   }
 
-  delete(vocs : Vocabulary[]) {
-    this.run.emit(new ops.DeleteVocabularies(vocs.map(v => v.id)))
+  delete(vocs: Vocabulary[]) {
+    this.run.emit(new ops.DeleteVocabularies(vocs.map((v) => v.id)));
     this.table.selection.clear();
   }
 
   addStandard() {
-    let vocIds = new Set(this.vocabularies.map(v => v.id));
-    this.api.vocabularies()
-      .pipe(map(vocs => {
-        return vocs.filter(v => !vocIds.has(v.id))
-      }))
-      .subscribe(vocs => {
+    let vocIds = new Set(this.vocabularies.map((v) => v.id));
+    this.api
+      .vocabularies()
+      .pipe(
+        map((vocs) => {
+          return vocs.filter((v) => !vocIds.has(v.id));
+        })
+      )
+      .subscribe((vocs) => {
         vocs.sort(Vocabulary.compare);
-        return this.dialog.open(VocabulariesDialogComponent, { data: { vocabularies: vocs } })
-          .afterClosed().subscribe(vocs => {
-            if (vocs != null) {
-              let cuis = Object.keys(this.mapping!.concepts);
-              let vocIds = (vocs as Vocabulary[]).map(v => v.id);
-              this.api.concepts(cuis, vocIds, this.mapping!.meta)
-                .subscribe(({ concepts, codes }) => {
-                  let conceptCodes =
-                    Object.fromEntries(
-                      Object.entries(concepts)
-                        .map(([cui, concept]) =>
-                          [cui, Object.fromEntries(
-                            Object.entries(concept.codes)
-                              .map(([vocId, codeIds]) => [vocId, Array.from(codeIds)]))]));
-                  let codes1 =
-                    Object.fromEntries(
-                      Object.entries(codes)
-                        .map(([vocId, codes]) => [vocId, Object.values(codes)]));
-                  this.run.emit(new ops.AddVocabularies(vocs, codes1, conceptCodes));
-                })
-            }
-          })
-      })
+        return this.dialog
+          .open(VocabulariesDialogComponent, { data: { vocabularies: vocs } })
+          .afterClosed()
+          .subscribe(async (vocs) => {
+            if (vocs == null) return;
+            let cuis = Object.keys(this.mapping!.concepts);
+            let vocIds = (vocs as Vocabulary[]).map((v) => v.id);
+            let { concepts, codes } = await this.api.concepts(
+              cuis,
+              vocIds,
+              this.mapping!.meta
+            );
+            let conceptCodes = Object.fromEntries(
+              Object.entries(concepts).map(([cui, concept]) => [
+                cui,
+                Object.fromEntries(
+                  Object.entries(concept.codes).map(([vocId, codeIds]) => [
+                    vocId,
+                    Array.from(codeIds),
+                  ])
+                ),
+              ])
+            );
+            let codes1 = Object.fromEntries(
+              Object.entries(codes).map(([vocId, codes]) => [
+                vocId,
+                Object.values(codes),
+              ])
+            );
+            this.run.emit(new ops.AddVocabularies(vocs, codes1, conceptCodes));
+          });
+      });
   }
 
   createCustom() {
-    this.dialog.open(CustomVocabularyDialogComponent, { data: { id: "", name: "", codesCSV: "" } })
-      .afterClosed().subscribe(data => {
-        let voc = new Vocabulary(data.id, data.name, null, true)
-        this.run.emit(new ops.AddVocabularies([voc], {}, {}));
+    this.dialog
+      .open(CustomVocabularyDialogComponent, {
+        data: { id: '', name: '', codesCSV: '' },
       })
+      .afterClosed()
+      .subscribe((data) => {
+        let voc = new Vocabulary(data.id, data.name, null, true);
+        this.run.emit(new ops.AddVocabularies([voc], {}, {}));
+      });
   }
 }